Created a rating component to render stars when necessary.
