+++ client/pages/bookmarks.jsx
@@ -1,36 +1,18 @@
 import React from 'react';
 import AppContext from '../../lib/app-context';
 import Redirect from '../components/redirect';
+import Rating from '../components/rating';
 
 export default class Bookmarks extends React.Component {
   constructor(props) {
     super(props);
     this.state = { bookmarks: [] };
     this.getBookmarks = this.getBookmarks.bind(this);
-    this.renderStars = this.renderStars.bind(this);
     this.visitBookmark = this.visitBookmark.bind(this);
   }
 
   componentDidMount() {
     this.getBookmarks();
-  }
-
-  renderStars(rating) {
-    const hasHalfStar = rating % 1 !== 0;
-    const fullStar = Math.floor(rating);
-    const emptyStar = Math.floor(5 - rating);
-    const arrayStars = [];
-    for (let i = 1; i <= fullStar; i++) {
-      arrayStars.push(<i key={`${i}-full-star`} className="fa-solid fa-star theme-color star-size "></i>);
-    }
-    if (hasHalfStar) {
-      arrayStars.push(<i key="1-half-star" className="fa-solid fa-star-half-stroke theme-color star-size "></i>);
-    }
-
-    for (let i = 1; i <= emptyStar; i++) {
-      arrayStars.push(<i key={`${i}-empty-star`} className="fa-regular fa-star theme-color star-size "></i>);
-    }
-    return arrayStars;
   }
 
   getBookmarks() {
@@ -73,7 +55,7 @@ export default class Bookmarks extends React.Component {
               <button onClick={this.visitBookmark} className='row border-radius bookmark align-center'>
                 <div className='column-two-thirds-nonresponsive'>
                   <p className='roboto-font result-info-size text-align-initial'>{business.name}</p>
-                  <div className='text-align-initial'>{this.renderStars(business.rating)}</div>
+                  <div className='text-align-initial'><Rating rating={business.rating}/></div>
                   <p className='roboto-font result-info-size text-align-initial margin-bottom-0'>{business.address1}</p>
                   {business.address2 && <p className='roboto-font result-info-size text-align-initial restaurant-info'>{business.address2}</p>}
                   <p className='roboto-font result-info-size text-align-initial restaurant-info'>{business.city} {business.state} {business.zipcode}</p>
